// modules/blocks-subgraph/generated/blocks-subgraph-types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    BigDecimal: string;
    BigInt: string;
    Bytes: string;
};

export type Block = {
    __typename?: 'Block';
    /** Author */
    author: Scalars['Bytes'];
    /** Difficulty */
    difficulty: Scalars['BigInt'];
    /** Gas Limit */
    gasLimit: Scalars['BigInt'];
    /** Gas Used */
    gasUsed: Scalars['BigInt'];
    /** ID (hash) */
    id: Scalars['ID'];
    /** Number */
    number: Scalars['BigInt'];
    /** Parent Hash */
    parentHash: Scalars['Bytes'];
    /** Receipts Root */
    receiptsRoot: Scalars['Bytes'];
    /** Size */
    size?: Maybe<Scalars['BigInt']>;
    /** State Root */
    stateRoot: Scalars['Bytes'];
    /** Timestamp */
    timestamp: Scalars['BigInt'];
    /** Total Difficulty */
    totalDifficulty: Scalars['BigInt'];
    /** Transactions Root */
    transactionsRoot: Scalars['Bytes'];
    /** Uncles Hash */
    unclesHash: Scalars['Bytes'];
};
// ...
export type BlockFragment = { __typename?: 'Block'; id: string; number: string; timestamp: string };