// modules/balancer-subgraph/generated/balancer-subgraph-types.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
    BigDecimal: string;
    BigInt: string;
    Bytes: string;
};
// ...
export type Balancer = {
    __typename?: 'Balancer';
    id: Scalars['ID'];
    poolCount: Scalars['Int'];
    pools?: Maybe<Array<Pool>>;
    totalLiquidity: Scalars['BigDecimal'];
    totalSwapCount: Scalars['BigInt'];
    totalSwapFee: Scalars['BigDecimal'];
    totalSwapVolume: Scalars['BigDecimal'];
};
// ...
export type GradualWeightUpdate = {
    __typename?: 'GradualWeightUpdate';
    endTimestamp: Scalars['Int'];
    endWeights: Array<Scalars['BigInt']>;
    id: Scalars['ID'];
    poolId: Pool;
    scheduledTimestamp: Scalars['Int'];
    startTimestamp: Scalars['Int'];
    startWeights: Array<Scalars['BigInt']>;
};
// ...
export type Investment = {
    __typename?: 'Investment';
    amount: Scalars['BigDecimal'];
    assetManagerAddress: Scalars['Bytes'];
    id: Scalars['ID'];
    poolTokenId: PoolToken;
    timestamp: Scalars['Int'];
};
// ...
export type Pool = {
    __typename?: 'Pool';
    address: Scalars['Bytes'];
    amp?: Maybe<Scalars['BigInt']>;
    baseToken?: Maybe<Scalars['Bytes']>;
    createTime: Scalars['Int'];
    expiryTime?: Maybe<Scalars['BigInt']>;
    factory?: Maybe<Scalars['Bytes']>;
    historicalValues?: Maybe<Array<PoolHistoricalLiquidity>>;
    holdersCount: Scalars['BigInt'];
    id: Scalars['ID'];
    managementFee?: Maybe<Scalars['BigDecimal']>;
    name?: Maybe<Scalars['String']>;
    owner?: Maybe<Scalars['Bytes']>;
    poolType?: Maybe<Scalars['String']>;
    priceRateProviders?: Maybe<Array<PriceRateProvider>>;
    principalToken?: Maybe<Scalars['Bytes']>;
    shares?: Maybe<Array<PoolShare>>;
    strategyType: Scalars['Int'];
    swapEnabled: Scalars['Boolean'];
    swapFee: Scalars['BigDecimal'];
    swaps?: Maybe<Array<Swap>>;
    swapsCount: Scalars['BigInt'];
    symbol?: Maybe<Scalars['String']>;
    tokens?: Maybe<Array<PoolToken>>;
    tokensList: Array<Scalars['Bytes']>;
    totalLiquidity: Scalars['BigDecimal'];
    totalShares: Scalars['BigDecimal'];
    totalSwapFee: Scalars['BigDecimal'];
    totalSwapVolume: Scalars['BigDecimal'];
    totalWeight?: Maybe<Scalars['BigDecimal']>;
    tx?: Maybe<Scalars['Bytes']>;
    unitSeconds?: Maybe<Scalars['BigInt']>;
    vaultID: Balancer;
    weightUpdates?: Maybe<Array<GradualWeightUpdate>>;
};
// ...
export type PoolHistoricalLiquidity = {
    __typename?: 'PoolHistoricalLiquidity';
    block: Scalars['BigInt'];
    id: Scalars['ID'];
    poolId: Pool;
    poolLiquidity: Scalars['BigDecimal'];
    poolLiquidityUSD: Scalars['BigDecimal'];
    poolShareValue: Scalars['BigDecimal'];
    poolTotalShares: Scalars['BigDecimal'];
    pricingAsset: Scalars['Bytes'];
    timestamp: Scalars['Int'];
};
// ...
export type PoolShare = {
    __typename?: 'PoolShare';
    balance: Scalars['BigDecimal'];
    id: Scalars['ID'];
    poolId: Pool;
    userAddress: User;
};
// ...
export type PoolToken = {
    __typename?: 'PoolToken';
    address: Scalars['String'];
    balance: Scalars['BigDecimal'];
    decimals: Scalars['Int'];
    id: Scalars['ID'];
    invested: Scalars['BigDecimal'];
    investments?: Maybe<Array<Investment>>;
    name: Scalars['String'];
    poolId: Pool;
    priceRate: Scalars['BigDecimal'];
    symbol: Scalars['String'];
    token: Token;
    weight?: Maybe<Scalars['BigDecimal']>;
};
// ...
export type PriceRateProvider = {
    __typename?: 'PriceRateProvider';
    address: Scalars['Bytes'];
    cacheDuration: Scalars['Int'];
    cacheExpiry: Scalars['Int'];
    id: Scalars['ID'];
    lastCached: Scalars['Int'];
    poolId: Pool;
    rate: Scalars['BigDecimal'];
    token: PoolToken;
};
// ...
export type Swap = {
    __typename?: 'Swap';
    caller: Scalars['Bytes'];
    id: Scalars['ID'];
    poolId: Pool;
    timestamp: Scalars['Int'];
    tokenAmountIn: Scalars['BigDecimal'];
    tokenAmountOut: Scalars['BigDecimal'];
    tokenIn: Scalars['Bytes'];
    tokenInSym: Scalars['String'];
    tokenOut: Scalars['Bytes'];
    tokenOutSym: Scalars['String'];
    tx: Scalars['Bytes'];
    userAddress: User;
    valueUSD: Scalars['BigDecimal'];
};
// ...
export type Token = {
    __typename?: 'Token';
    address: Scalars['String'];
    decimals: Scalars['Int'];
    id: Scalars['ID'];
    latestPrice?: Maybe<LatestPrice>;
    name?: Maybe<Scalars['String']>;
    symbol?: Maybe<Scalars['String']>;
    totalBalanceNotional: Scalars['BigDecimal'];
    totalBalanceUSD: Scalars['BigDecimal'];
    totalSwapCount: Scalars['BigInt'];
    totalVolumeNotional: Scalars['BigDecimal'];
    totalVolumeUSD: Scalars['BigDecimal'];
};
// ...
export type User = {
    __typename?: 'User';
    id: Scalars['ID'];
    sharesOwned?: Maybe<Array<PoolShare>>;
    swaps?: Maybe<Array<Swap>>;
    userInternalBalances?: Maybe<Array<UserInternalBalance>>;
};
// ...
export type UserInternalBalance = {
    __typename?: 'UserInternalBalance';
    balance: Scalars['BigDecimal'];
    id: Scalars['ID'];
    token: Scalars['Bytes'];
    userAddress?: Maybe<User>;
};
// ...
export type BalancerUserFragment = {
    __typename?: 'User';
    id: string;
    sharesOwned?: Array<{
        __typename?: 'PoolShare';
        id: string;
        balance: string;
        poolId: { __typename?: 'Pool'; id: string };
    }> | null;
};
// ...
export type BalancerPoolFragment = {
    __typename?: 'Pool';
    id: string;
    address: string;
    poolType?: string | null;
    symbol?: string | null;
    name?: string | null;
    swapFee: string;
    totalWeight?: string | null;
    totalSwapVolume: string;
    totalSwapFee: string;
    totalLiquidity: string;
    totalShares: string;
    swapsCount: string;
    holdersCount: string;
    createTime: number;
    swapEnabled: boolean;
    tokensList: Array<string>;
    factory?: string | null;
    expiryTime?: string | null;
    unitSeconds?: string | null;
    principalToken?: string | null;
    baseToken?: string | null;
    owner?: string | null;
    amp?: string | null;
    tokens?: Array<{
        __typename?: 'PoolToken';
        id: string;
        symbol: string;
        name: string;
        decimals: number;
        address: string;
        balance: string;
        invested: string;
        weight?: string | null;
        priceRate: string;
    }> | null;
};